risk_api/requirements.txt
fastapi==0.104.1
gunicorn==21.2.0
httpx==0.25.1
loguru==0.7.2
motor==3.3.1
orjson==3.9.10
pydantic==2.4.2
pydantic_settings==2.0.3
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
uvicorn==0.24.0.post1
